added attr_protected calls to protect mass assignment of :project_id through a view

+++ app/models/member_role.rb
@@ -21,6 +21,8 @@ class MemberRole < ActiveRecord::Base
   after_create :add_role_to_group_users
   after_destroy :remove_role_from_group_users
 
+  #attr_protected :member_id #, :role_id
+
   validates_presence_of :role
 
   def validate
@@ -34,7 +36,7 @@ class MemberRole < ActiveRecord::Base
   private
 
   def remove_member_if_empty
-    if member.roles.empty?
+    if member && member.roles.empty?
       member.destroy
     end
   end
